fix(core): ensure daemon enabled check is unchanged
