+++ packages/nx/src/daemon/client/client.ts
@@ -16,6 +16,7 @@ import { getFullOsSocketPath, killSocketOrPath } from '../socket-utils';
 import {
   DAEMON_DIR_FOR_CURRENT_WORKSPACE,
   DAEMON_OUTPUT_LOG_FILE,
+  isDaemonDisabled,
   removeSocketDir,
 } from '../tmp-dir';
 import { FileData, ProjectGraph } from '../../config/project-graph';
@@ -95,7 +96,6 @@ import {
   POST_TASKS_EXECUTION,
   PRE_TASKS_EXECUTION,
 } from '../message-types/run-tasks-execution-hooks';
-import { isDaemonEnabled } from './enabled';
 
 const DAEMON_ENV_SETTINGS = {
   NX_PROJECT_GLOB_CACHE: 'false',
@@ -141,7 +141,48 @@ export class DaemonClient {
   private _err: FileHandle = null;
 
   enabled() {
-    return isDaemonEnabled(this.nxJson);
+    if (this._enabled === undefined) {
+      const useDaemonProcessOption = this.nxJson?.useDaemonProcess;
+      const env = process.env.NX_DAEMON;
+
+      // env takes precedence
+      // option=true,env=false => no daemon
+      // option=false,env=undefined => no daemon
+      // option=false,env=false => no daemon
+
+      // option=undefined,env=undefined => daemon
+      // option=true,env=true => daemon
+      // option=false,env=true => daemon
+
+      // CI=true,env=undefined => no daemon
+      // CI=true,env=false => no daemon
+      // CI=true,env=true => daemon
+
+      // docker=true,env=undefined => no daemon
+      // docker=true,env=false => no daemon
+      // docker=true,env=true => daemon
+      // WASM => no daemon because file watching does not work
+      if (
+        ((isCI() || isDocker()) && env !== 'true') ||
+        isDaemonDisabled() ||
+        nxJsonIsNotPresent() ||
+        (useDaemonProcessOption === undefined && env === 'false') ||
+        (useDaemonProcessOption === true && env === 'false') ||
+        (useDaemonProcessOption === false && env === undefined) ||
+        (useDaemonProcessOption === false && env === 'false')
+      ) {
+        this._enabled = false;
+      } else if (IS_WASM) {
+        output.warn({
+          title:
+            'The Nx Daemon is unsupported in WebAssembly environments. Some things may be slower than or not function as expected.',
+        });
+        this._enabled = false;
+      } else {
+        this._enabled = true;
+      }
+    }
+    return this._enabled;
   }
 
   reset() {
@@ -676,6 +717,27 @@ export class DaemonClient {
 }
 
 export const daemonClient = new DaemonClient();
+
+export function isDaemonEnabled() {
+  return daemonClient.enabled();
+}
+
+function isDocker() {
+  try {
+    statSync('/.dockerenv');
+    return true;
+  } catch {
+    try {
+      return readFileSync('/proc/self/cgroup', 'utf8')?.includes('docker');
+    } catch {}
+
+    return false;
+  }
+}
+
+function nxJsonIsNotPresent() {
+  return !hasNxJson(workspaceRoot);
+}
 
 function daemonProcessException(message: string) {
   try {
