+++ packages/nx/src/devkit-exports.ts
@@ -260,4 +260,4 @@ export { cacheDir } from './utils/cache-directory';
  */
 export { createProjectFileMapUsingProjectGraph } from './project-graph/file-map-utils';
 
-export { isDaemonEnabled } from './daemon/client/enabled';
+export { isDaemonEnabled } from './daemon/client/client';
