+++ packages/rspack/src/plugins/utils/plugins/rspack-nx-build-coordination-plugin.ts
@@ -1,7 +1,6 @@
 import { exec } from 'child_process';
 import type { Compiler } from '@rspack/core';
-import { daemonClient } from 'nx/src/daemon/client/client';
-import { isDaemonEnabled } from 'nx/src/daemon/client/enabled';
+import { daemonClient, isDaemonEnabled } from 'nx/src/daemon/client/client';
 import { BatchFunctionRunner } from 'nx/src/command-line/watch/watch';
 import { output } from 'nx/src/utils/output';
